backend fully adjusted

+++ src/utils/Cloudinary.js
@@ -2,7 +2,6 @@
 import { v2 as cloudinary } from 'cloudinary';
 import { extractPublicId } from 'cloudinary-build-url';
 import fs from 'fs';
-import path from 'path';
 
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -10,26 +9,25 @@ cloudinary.config({
     api_secret: process.env.CLOUDINARY_SECRET_KEY,
 });
 
-console.log(`Cloudinary Config:
-    Cloud Name: ${process.env.CLOUDINARY_CLOUD_NAME},
-    API Key: ${process.env.CLOUDINARY_API_KEY},
-    API Secret: ${process.env.CLOUDINARY_SECRET_KEY ? '********SET' : 'Not Set'}`);
-
 
 const uploadImageToCloudinary = async (filePath) => {
     try {
-        console.log(`Starting upload for: ${filePath}`);
+        if (!filePath) return null;
 
         const uploadedFile = await cloudinary.uploader.upload(filePath, { resource_type: 'auto' });
 
         console.log(`Upload successful: ${uploadedFile.url}`);
 
-        fs.unlinkSync(filePath); // Remove file from disk after upload
+        fs.unlinkSync(filePath);
+
         return uploadedFile;
+
     } catch (error) {
         console.error(`Cloudinary File Uploading Error ==> ${error.message}`);
+
         if (fs.existsSync(filePath)) {
-            fs.unlinkSync(filePath); // Ensure file is removed in case of error
+
+            fs.unlinkSync(filePath);
         }
         return null;
     }
@@ -37,11 +35,15 @@ const uploadImageToCloudinary = async (filePath) => {
 
 const deleteImageFromCloudinary = async (fileUrl) => {
     const filePath = extractPublicId(fileUrl);
+
     try {
-        const deletedFile = await cloudinary.uploader.destroy(filePath, { resource_type: 'image' });
+        const deletedFile = await cloudinary.uploader
+            .destroy(filePath, { resource_type: 'image' });
+
         if (deletedFile.result === 'ok') {
             console.log(`The File is deleted Successfully.`);
         }
+
         return deletedFile;
     } catch (error) {
         console.log(`Cloudinary File Deleting Error ==> ${error.message}`);
